requirements.txt
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
clowncar==0.2.0
DateTime==4.3
Django==3.1.3
fsspec==0.8.5
groupy==0.5.2
GroupyAPI==0.10.3
idna==2.7
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
prettyprint==0.1.5
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.20.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typing==3.7.4.3
tz==0.2.2
urllib3==1.24.3
wrapt==1.12.1
zope.interface==5.2.0
